Rename GithubMarkdownPlugin to GithubPlugin. Remove redundant component exports and mark sources for refactoring.
