+++ packages/apps/composer-app/src/plugins/GithubPlugin/components/_EmbeddedMain/embedded.css
