+++ packages/apps/composer-app/src/plugins/GithubPlugin/components/_StandaloneMain/index.ts
