+++ packages/apps/composer-app/src/plugins/GithubPlugin/components/index.ts
@@ -4,5 +4,3 @@
 
 export * from './GithubApiProviders';
 export * from './GithubEchoResolverProviders';
-export * from './StandaloneMain';
-export * from './EmbeddedMain';
