+++ packages/apps/composer-app/src/plugins/GithubPlugin/index.ts
@@ -3,5 +3,3 @@
 //
 
 export * from './GithubPlugin';
-export * from './MarkdownPlugin';
-export * from './SpacePlugin';
